Update Fontbakery and profile definitions
